Allow RowsMetadata with empty column definitions (fixes #9)
Motivation:

Currently we always set NO_METADATA when the column list is empty. This
is wrong, the server can legitimately return a response with no column
specs but NO_METADATA == false, for example when preparing a statement
that has no variables.

Modifications:

Add a specific constructor for the case where NO_METADATA should be set.
The other constructor that takes a list of column definitions never sets
NO_METADATA (even if the list is empty).

Result:

It's now possible to create a response that doesn't skip metadata, but
also has zero columns.